requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
crispy-bootstrap4==2024.1
cymem==2.0.8
datasets==3.1.0
dill==0.3.8
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
djangorestframework==3.15.2
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.26.2
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jiter==0.6.0
joblib==1.4.2
langcodes==3.4.1
langdetect==1.0.9
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.4.2
numpy==2.0.2
openai==0.28.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
preshed==3.0.9
psutil==6.1.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.66.5
transformers==4.46.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
